refacto(AnswerCampagne): move functions to dedicated utils file
